# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather & Auth API",
        "description": "Documentation for the Weather & Auth API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8004"
        },
        {
            "url": "http://ms-weather.sytes.net"
        }
    ],
    "paths": {
...
        "/api/v1/weather-module/favorite-cities/{id}": {
            "delete": {
                "tags": [
                    "Favorite Cities"
                ],
                "summary": "Remove a city from favorites",
                "operationId": "37a45b5546b93a48d96d6638a05487a1",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language for API responses (en or es)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "es"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the favorite city to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City removed from favorites successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "City removed from favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Favorite city not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Favorite city not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Error message from the service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}